Finish, refactor, and improve drag to reorder functionality

+++ app/src/main/java/com/rendox/grocerygenius/database/grocery_list/GroceryListEntity.kt
@@ -7,5 +7,5 @@ import androidx.room.PrimaryKey
 data class GroceryListEntity(
     @PrimaryKey val id: String,
     val name: String,
-    val sortingPriority: Int,
+    val sortingPriority: Long,
 )
